URL定義にしたがって現在の構造やリンクを書き換えた
