+++ service/pages/index.tsx
@@ -1,43 +1,20 @@
-import { GetStaticProps, InferGetStaticPropsType } from 'next'
-import { useState } from 'react'
-import { SummaryBookPokemon } from '../types/pokemon/SummaryBookPokemon'
-
 import Header from '../components/common/Header';
-import PageTitle from '../components/common/PageTitle';
-import BookPokemonContents from '../components/bookpokemon/Contents';
 import Footer from '../components/common/Footer';
-import getSummaryBookPokemonList from '../lib/pokemon/getSummaryBookPokemonList';
+import styled from 'styled-components';
 
-export default function Home({ initialPokemons, isNext }: InferGetStaticPropsType<typeof getStaticProps>) {
-
-  const [pager, setPager] = useState<number>(1);
-  const [next, setNext] = useState<boolean>(isNext);
-  const [pokemons, setPokemons] = useState<SummaryBookPokemon[]>(initialPokemons);
-
-  const more = async () => {
-    setPager(pager + 1);
-    const { result, isNext } = await getSummaryBookPokemonList(pager * 20, 20);
-    setPokemons(pokemons.concat(result));
-    setNext(isNext);
-  }
+export default function Home() {
 
   return (
     <>
       <Header />
-      <PageTitle title='ポケモン図鑑そうごう' menu='book' />
-      <BookPokemonContents next={next} pokemons={pokemons} pager={pager} viewMore={more}  />
+        <Title>ポケモンを探そう</Title>
       <Footer />
     </>
   )  
 }
 
-export const getStaticProps: GetStaticProps = async () => {
-
-  const { result, isNext } = await getSummaryBookPokemonList();
-  return {
-    props: {
-      initialPokemons: result,
-      isNext
-    }
-  };
-}
+export const Title = styled.p`
+  margin: 200px auto;
+  text-align: center;
+  font-size: 32px;
+`;
